[Modify]<src/components/Modal/...> separate component as file
각 컴포넌트들을 modal wrapper안에서 랜더링하는것이 아닌, 각자 modal wrapper를 이용해 별개의 컴포넌트로 사용하기 위해 파일 단위로 분리
