+++ src/components/Modals/BugReport/BugReport.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 // import ModalCloseImg from '../../assets/icon/black-close-button.png';
 // import BugIconImg from '../../assets/icon/bug.png';
 
-import * as S from './style';
+import * as S from '../style';
 
 interface Props {}
