+++ src/components/Modals/Login/Login.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import CloseImg from '../../assets/icon/black-close-button.png';
-import * as S from './style';
+import * as S from '../style';
 
 interface Props {}
